requirements.txt

google-generativeai == 2024.2.2